Latest attempt at getting the timer to press on is more of a hack than anything, but it should work.  Best to wait and see on that one first, though.

+++ tabata_timer.py
@@ -6,9 +6,6 @@ class TabataTimer(object):
 	def __init__(self):
 		super(TabataTimer, self).__init__()
 
-	def __del___(self):
-		self.stop()		
-
 	def _start(self):
 		self._proc = subprocess.Popen(['python3', 'base_timer.py'],
                             stdout=subprocess.PIPE,
@@ -21,16 +18,22 @@ class TabataTimer(object):
 			self._start()
 
 	def stop(self):
-		self._proc.terminate()
-		self._proc.wait()
-		_ = self.getReturnCode()
+		if '_proc' in dir(self):
+			self._proc.terminate()
+			self._proc.wait()
+			_ = self.getReturnCode()
 
 	def getPid(self):
 		if '_proc' in dir(self):
 			return self._proc.pid
 
 	def pulse(self):
-		self._proc.send_signal(0)
+		try:
+			self._proc.communicate(timeout=0.01)
+		except subprocess.TimeoutExpired:
+			pass
+		finally:
+			return 1			
 
 	def getReturnCode(self):
 		_outs, _ = self._proc.communicate(timeout=0.2)
